+++ ash-ra-template/src/vivid/art/evaluate.clj
@@ -5,18 +5,25 @@
     [clojure.string]
     [vivid.art.embed :as embed]))
 
-(defn wrap-forms
-  [forms]
-  (concat ["(ns user)"
-           "(def ^java.lang.StringBuilder __vivid__art__sb (new java.lang.StringBuilder))"
-           "(defn emit [val] (.append user/__vivid__art__sb val))"]
-          forms
-          ["(.toString user/__vivid__art__sb)"]))
+(def ^:const prelude ["(ns user)"])
+(def ^:const interlude ["(def ^java.lang.StringBuilder __vivid__art__sb (new java.lang.StringBuilder))"
+                        "(defn emit ([]) ([v] (.append user/__vivid__art__sb v) nil))"])
+(def ^:const coda ["(.toString user/__vivid__art__sb)"])
+
+(defn define-bindings
+  [bindings]
+  (vec (for [[k v] bindings] (format "(def %s %s)"
+                                     (pr-str k)
+                                     (pr-str v)))))
 
 (defn evaluate
-  [forms
-   & {:keys [dependencies]}]
-  (let [wrapped-forms (wrap-forms forms)
-        str (clojure.string/join "\n" wrapped-forms)]
-    (embed/eval-in-one-shot-runtime str
-                                    :dependencies dependencies)))
+  [forms bindings dependencies]
+  (let [code (->> (interpose [""]
+                             [prelude
+                              (define-bindings bindings)
+                              interlude
+                              forms
+                              coda])
+                  (flatten)
+                  (clojure.string/join "\n"))]
+    (embed/eval-in-one-shot-runtime code dependencies)))