+++ lein-art/resources/test-project.clj
@@ -1,0 +1,6 @@
+; Copyright 2019 Vivid Inc.
+
+(defproject lein-art-test "0.0.0"
+
+            :dependencies [[org.clojure/clojure "1.9.0"]
+                           [vivid/ash-ra-template "0.4.0"]])